% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_arm_names}
\alias{get_arm_names}
\title{Getter for treatment arm names}
\usage{
get_arm_names(input)
}
\arguments{
\item{input}{A list of parameters or model results.}
}
\value{
A character vector of treatment arm names.
}
\description{
Getter for treatment arm names
}
\examples{
data(test_data)
get_arm_names(test_data)
}
