% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_costs}
\alias{get_costs}
\title{Getter for costs}
\usage{
get_costs(res)
}
\arguments{
\item{res}{A list of results returned by ce_markov.}
}
\value{
A numeric vector of total costs by treatment.
}
\description{
Getter for costs
}
\examples{
data(test_data)
res <- run_model(test_data)
get_costs(res)
}
