% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_icer}
\alias{get_icer}
\title{Getter for ICER}
\usage{
get_icer(res, arm1 = "without_drug", arm2 = "with_drug")
}
\arguments{
\item{res}{A list of results returned by ce_markov.}

\item{arm1}{Baseline treatment arm name or index (default: "without_drug").}

\item{arm2}{Comparator treatment arm name or index (default: "with_drug").}
}
\value{
A numeric value representing the incremental cost-effectiveness ratio (ICER).
}
\description{
Getter for ICER
}
\examples{
data(test_data)
res <- run_model(test_data)
get_icer(res)
}
