% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_incremental_costs}
\alias{get_incremental_costs}
\title{Getter for incremental costs}
\usage{
get_incremental_costs(res, arm1 = "without_drug", arm2 = "with_drug")
}
\arguments{
\item{res}{A list of results returned by ce_markov.}

\item{arm1}{Baseline treatment arm name or index (default: "without_drug").}

\item{arm2}{Comparator treatment arm name or index (default: "with_drug").}
}
\value{
A numeric value of incremental cost (arm 2 minus arm 1).
}
\description{
Getter for incremental costs
}
\examples{
data(test_data)
res <- run_model(test_data)
get_incremental_costs(res)
}
