% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_transition_matrix}
\alias{get_transition_matrix}
\title{Getter for transition probability matrix}
\usage{
get_transition_matrix(input, arm = NULL)
}
\arguments{
\item{input}{A list of model parameters.}

\item{arm}{Optional treatment arm name (character) or index.}
}
\value{
A matrix of transition probabilities for the specified arm, or the 3D array if arm is NULL.
}
\description{
Getter for transition probability matrix
}
\examples{
data(test_data)
get_transition_matrix(test_data, "without_drug")
}
