% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_helpers.R
\name{run_model}
\alias{run_model}
\title{Run Markov model with parameters and overrides}
\usage{
run_model(data, ...)
}
\arguments{
\item{data}{A list of base parameters.}

\item{...}{Overrides for parameters or helper arguments.}
}
\value{
A list of model simulation results.
}
\description{
Run Markov model with parameters and overrides
}
\examples{
data(test_data)
res <- run_model(test_data, discount_rate = 0.05)
}
