% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_costs}
\alias{set_costs}
\title{Set costs for a scenario}
\usage{
set_costs(input, scenario, values)
}
\arguments{
\item{input}{A list of model parameters.}

\item{scenario}{The scenario name (character) or index (numeric).}

\item{values}{A vector of cost values for health states.}
}
\value{
The modified list of parameters.
}
\description{
Set costs for a scenario
}
\examples{
data(test_data)
updated <- set_costs(test_data, "without_drug", c(100, 200, 0))
}
