% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_discount_rate}
\alias{set_discount_rate}
\title{Set discount rate}
\usage{
set_discount_rate(input, discount_rate)
}
\arguments{
\item{input}{A list of model parameters.}

\item{discount_rate}{The discount rate to set (numeric).}
}
\value{
The modified list of parameters.
}
\description{
Set discount rate
}
\examples{
data(test_data)
updated <- set_discount_rate(test_data, 0.05)
}
