% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_p_matrix}
\alias{set_p_matrix}
\title{Set transition probabilities for a scenario}
\usage{
set_p_matrix(input, scenario, values)
}
\arguments{
\item{input}{A list of model parameters.}

\item{scenario}{The scenario name (character) or index (numeric).}

\item{values}{A matrix of transition probabilities.}
}
\value{
The modified list of parameters.
}
\description{
Set transition probabilities for a scenario
}
\examples{
data(test_data)
mat <- test_data$p_matrix[, , "without_drug"]
updated <- set_p_matrix(test_data, "without_drug", mat)
}
