% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_time_horizon}
\alias{set_time_horizon}
\title{Set time horizon}
\usage{
set_time_horizon(input, scenario, values)
}
\arguments{
\item{input}{A list of model parameters.}

\item{scenario}{The scenario name or index, or values if scenario is omitted.}

\item{values}{The number of cycles (numeric).}
}
\value{
The modified list of parameters.
}
\description{
Set time horizon
}
\examples{
data(test_data)
updated <- set_time_horizon(test_data, 10)
}
