% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{swap_arm_params}
\alias{swap_arm_params}
\title{Swap treatment-specific parameters between two arms}
\usage{
swap_arm_params(input, arm1 = "without_drug", arm2 = "with_drug")
}
\arguments{
\item{input}{A list of model parameters.}

\item{arm1}{Name or index of the first treatment arm.}

\item{arm2}{Name or index of the second treatment arm.}
}
\value{
The modified list of parameters with swapped inputs.
}
\description{
Swap treatment-specific parameters between two arms
}
\examples{
data(test_data)
updated <- swap_arm_params(test_data, "without_drug", "with_drug")
}
