% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_matrix_cycle.R
\name{tpDn_lookup}
\alias{tpDn_lookup}
\title{Look-up table for age-specific probabilities of death from non-disease causes}
\usage{
tpDn_lookup()
}
\description{
Look-up table for age-specific probabilities of death from non-disease causes
}
