% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_markov.R
\name{ce_markov}
\alias{ce_markov}
\title{run cost-effectiveness model}
\usage{
ce_markov(
  start_pop,
  p_matrix,
  state_c_matrix,
  trans_c_matrix,
  state_q_matrix,
  n_cycles = 46,
  init_age = 55,
  s_names = NULL,
  t_names = NULL,
  discount_rate = 0.035,
  ...
)
}
\arguments{
\item{start_pop}{A vector of initial population counts for each health state}

\item{p_matrix}{A 3D array of transition probabilities for each treatment}

\item{state_c_matrix}{A matrix of state costs for each treatment}

\item{trans_c_matrix}{A matrix of transition costs for each treatment}

\item{state_q_matrix}{A matrix of state utilities for each treatment}

\item{n_cycles}{The number of cycles to run the model for}

\item{init_age}{The initial age of the population}

\item{s_names}{Optional names for the health states}

\item{t_names}{Optional names for the treatments}

\item{discount_rate}{The discount rate to apply to costs and QALYs}

\item{...}{Additional arguments passed to p_matrix_cycle.}
}
\value{
A list containing the population array, total life expectancy, total life-years,
cycle costs, cycle QALYs, total costs, and total QALYs for each treatment
}
\description{
Runs a Markov model to calculate costs, QALYs, and life expectancy for different treatments
}
