% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_helpers.R
\name{check_model_costs}
\alias{check_model_costs}
\title{Run model and check Costs against an expected value}
\usage{
check_model_costs(expected_costs, ..., label = NULL)
}
\arguments{
\item{expected_costs}{The exact expected Cost value.}

\item{...}{Arguments passed to \code{run_model()}.}

\item{label}{A label for the test, passed to \code{expect_equal()}.}
}
\description{
Run model and check Costs against an expected value
}
