% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_helpers.R
\name{check_model_le}
\alias{check_model_le}
\title{Run model and check Life Expectancy against an expected value}
\usage{
check_model_le(expected_le, ..., label = NULL)
}
\arguments{
\item{expected_le}{The exact expected LE value.}

\item{...}{Arguments passed to \code{run_model()}.}

\item{label}{A label for the test, passed to \code{expect_equal()}.}
}
\description{
Run model and check Life Expectancy against an expected value
}
