% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_incremental_qalys}
\alias{get_incremental_qalys}
\title{Getter for incremental QALYs}
\usage{
get_incremental_qalys(res)
}
\arguments{
\item{res}{A list of results returned by ce_markov.}
}
\value{
A numeric value of incremental QALYs (arm 2 minus arm 1).
}
\description{
Getter for incremental QALYs
}
\examples{
data(test_data)
res <- run_model(test_data)
get_incremental_qalys(res)
}
