% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_le}
\alias{get_le}
\title{Getter for life expectancy (LE)}
\usage{
get_le(res, arm = NULL)
}
\arguments{
\item{res}{A list of results returned by ce_markov.}

\item{arm}{Optional treatment arm name or index.}
}
\value{
A numeric vector of total life expectancy by treatment, or a single numeric value if arm is specified.
}
\description{
Getter for life expectancy (LE)
}
\examples{
data(test_data)
res <- run_model(test_data)
get_le(res)
get_le(res, "with_drug")
}
