% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_qalys}
\alias{get_qalys}
\title{Getter for QALYs}
\usage{
get_qalys(res, arm = NULL)
}
\arguments{
\item{res}{A list of results returned by ce_markov.}

\item{arm}{Optional treatment arm name or index.}
}
\value{
A numeric vector of total QALYs by treatment, or a single numeric value if arm is specified.
}
\description{
Getter for QALYs
}
\examples{
data(test_data)
res <- run_model(test_data)
get_qalys(res)
get_qalys(res, "with_drug")
}
