% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_utilities}
\alias{set_utilities}
\title{Set utilities for a scenario}
\usage{
set_utilities(input, scenario, values)
}
\arguments{
\item{input}{A list of model parameters.}

\item{scenario}{The scenario name (character) or index (numeric).}

\item{values}{A vector of utility values for health states.}
}
\value{
The modified list of parameters.
}
\description{
Set utilities for a scenario
}
